workstation_automations/python_scripts/piper_tts_env/requirements.txt
piper-tts
PyPDF2
